# Conflicts:
#	pom.xml

+++ src/main/java/com/shopping/application/controller/UserController.java
@@ -7,11 +7,13 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.shopping.application.dto.UserDto;
+import com.shopping.application.exception.UserNotFoundException;
 import com.shopping.application.mapper.UserMapper;
 import com.shopping.application.models.User;
 import com.shopping.application.service.UserService;
@@ -33,11 +35,6 @@ public class UserController {
         return ResponseEntity.ok(userMapper.usersToUserDtos(userService.getAll()));
     }
     
-    @GetMapping("/{id}")
-    ResponseEntity<UserDto> getById(@PathVariable String id){
-        return ResponseEntity.ok(userMapper.userToUserDto(userService.getById(id)));
-    }
-    
     @PostMapping
     ResponseEntity<UserDto> create(@RequestBody UserDto userDto){
         
@@ -45,8 +42,25 @@ public class UserController {
         return ResponseEntity.status(HttpStatus.CREATED).body(userMapper.userToUserDto(userService.createUser(user)));
     }
     
-    @DeleteMapping
-    ResponseEntity<?> delete(@RequestBody UserDto userDto){
+    @GetMapping("/{id}")
+    ResponseEntity<UserDto> getById(@PathVariable String id) throws UserNotFoundException {
+        return ResponseEntity.ok(userMapper.userToUserDto(userService.getById(id)));
+    }
+    
+    @PutMapping("/{id}")
+    ResponseEntity<?> update(@PathVariable String id, @RequestBody UserDto userDto) throws UserNotFoundException {
+        if(!id.equals(userDto.getId())) {
+            return ResponseEntity.badRequest().body("The Id in parameter must be the same in the body of the request");
+        }
+        User user = userMapper.userDtoToUser(userDto);
+        return ResponseEntity.ok(userMapper.userToUserDto(userService.updateUser(user)));
+    }
+        
+    @DeleteMapping("/{id}")
+    ResponseEntity<?> delete(@PathVariable String id, @RequestBody UserDto userDto) throws UserNotFoundException {
+        if(!id.equals(userDto.getId())) {
+            return ResponseEntity.badRequest().body("The Id in parameter must be the same in the body of the request");
+        }
         userService.deleteUser(userMapper.userDtoToUser(userDto));
         return ResponseEntity.ok().build();
     }
